rename changed snapgrid package name as it was wrongly written
